requirements.txt
Pillow>=9.3
requests
python-dateutil
